Sb3 integration improvements (#2418)
* push changes

* quick fix

* fix reformat

* use a logger

* improves logging
