+++ functional_tests/sb3-integration.py
@@ -15,8 +15,6 @@ assert:
   - :wandb:runs[0][exitcode]: 0
 """
 
-import time
-
 import gym
 from stable_baselines3 import PPO
 from stable_baselines3.common.monitor import Monitor
@@ -24,11 +22,12 @@ from stable_baselines3.common.vec_env import DummyVecEnv
 import wandb
 from wandb.integration.sb3 import WandbCallback
 
-
-config = {"policy_type": "MlpPolicy", "total_timesteps": 200}
-experiment_name = f"PPO_{int(time.time())}"
+config = {
+    "policy_type": "MlpPolicy",
+    "total_timesteps": 200,
+    "env_name": "CartPole-v1",
+}
 run = wandb.init(
-    name=experiment_name,
     project="sb3",
     config=config,
     sync_tensorboard=True,  # auto-upload sb3's tensorboard metrics
@@ -37,21 +36,20 @@ run = wandb.init(
 
 
 def make_env():
-    env = gym.make("CartPole-v1")
+    env = gym.make(config["env_name"])
     env = Monitor(env)  # record stats such as returns
     return env
 
 
 env = DummyVecEnv([make_env])
 model = PPO(
-    config["policy_type"], env, verbose=1, tensorboard_log=f"runs/{experiment_name}"
+    config["policy_type"], env, verbose=1, tensorboard_log=f"runs/{run.name}"
 )
 
 model.learn(
     total_timesteps=config["total_timesteps"],
     callback=WandbCallback(
         gradient_save_freq=100,
-        model_save_freq=1000,
-        model_save_path=f"models/{experiment_name}",
+        model_save_path=f"models/{run.name}",
     ),
 )
